+++ src/test/config_test.ts
@@ -60,7 +60,9 @@ suite('makeConfig', function() {
               width: 1024,
             },
           },
-          measurement: 'callback',
+          measurement: {
+            kind: 'callback',
+          },
           name: 'random-global.html',
           url: {
             kind: 'local',
@@ -101,7 +103,9 @@ suite('makeConfig', function() {
               width: 1024,
             },
           },
-          measurement: 'callback',
+          measurement: {
+            kind: 'callback',
+          },
           // TODO(aomarks) Why does this have a forward-slash?
           name: '/random-global.html',
           url: {
@@ -142,7 +146,9 @@ suite('makeConfig', function() {
               width: 1024,
             },
           },
-          measurement: 'callback',
+          measurement: {
+            kind: 'callback',
+          },
           // TODO(aomarks) Why does this have a forward-slash?
           name: '/random-global.html',
           url: {
@@ -190,7 +196,9 @@ suite('makeConfig', function() {
               width: 1024,
             },
           },
-          measurement: 'callback',
+          measurement: {
+            kind: 'callback',
+          },
           // TODO(aomarks) Why does this have a forward-slash?
           name: '/random-global.html',
           url: {